Enum blunder::BsdError

pub enum BsdError {
    EPERM,
    ENOENT,
    ESRCH,
    EINTR,
    EIO,
    ENXIO,
    E2BIG,
    ENOEXEC,
    EBADF,
    ECHILD,
    EDEADLK,
    ENOMEM,
    EACCES,
    EFAULT,
    ENOTBLK,
    EBUSY,
    EEXIST,
    EXDEV,
    ENODEV,
    ENOTDIR,
    EISDIR,
    EINVAL,
    ENFILE,
    EMFILE,
    ENOTTY,
    ETXTBSY,
    EFBIG,
    ENOSPC,
    ESPIPE,
    EROFS,
    EMLINK,
    EPIPE,
    EDOM,
    ERANGE,
    EAGAIN,
    EINPROGRESS,
    EALREADY,
    ENOTSOCK,
    EDESTADDRREQ,
    EMSGSIZE,
    EPROTOTYPE,
    ENOPROTOOPT,
    EPROTONOSUPPORT,
    ESOCKTNOSUPPORT,
    EOPNOTSUPP,
    EPFNOSUPPORT,
    EAFNOSUPPORT,
    EADDRINUSE,
    EADDRNOTAVAIL,
    ENETDOWN,
    ENETUNREACH,
    ENETRESET,
    ECONNABORTED,
    ECONNRESET,
    ENOBUFS,
    EISCONN,
    ENOTCONN,
    ESHUTDOWN,
    ETIMEDOUT,
    ECONNREFUSED,
    ELOOP,
    ENAMETOOLONG,
    EHOSTDOWN,
    EHOSTUNREACH,
    ENOTEMPTY,
    EPROCLIM,
    EUSERS,
    EDQUOT,
    ESTALE,
    EBADRPC,
    ERPCMISMATCH,
    EPROGUNAVAIL,
    EPROGMISMATCH,
    EPROCUNAVAIL,
    ENOLCK,
    ENOSYS,
    EFTYPE,
    EAUTH,
    ENEEDAUTH,
    EIDRM,
    ENOMSG,
    EOVERFLOW,
    ECANCELED,
    EILSEQ,
    ENOATTR,
    EDOOFUS,
    EBADMSG,
    EMULTIHOP,
    ENOLINK,
    EPROTO,
    ENOTCAPABLE,
    ECAPMODE,
    ENOTRECOVERABLE,
    EOWNERDEAD,
}

Errors that can be encoutered while working with FreeBSD's libc. Some of them are BSD specific, some of them are from POSIX.

Variants

EPERM

ENOENT

ESRCH

EINTR

EIO

ENXIO

E2BIG

ENOEXEC

EBADF

ECHILD

EDEADLK

ENOMEM

EACCES

EFAULT

ENOTBLK

EBUSY

EEXIST

EXDEV

ENODEV

ENOTDIR

EISDIR

EINVAL

ENFILE

EMFILE

ENOTTY

ETXTBSY

EFBIG

ENOSPC

ESPIPE

EROFS

EMLINK

EPIPE

EDOM

ERANGE

EAGAIN

EINPROGRESS

EALREADY

ENOTSOCK

EDESTADDRREQ

EMSGSIZE

EPROTOTYPE

ENOPROTOOPT

EPROTONOSUPPORT

ESOCKTNOSUPPORT

EOPNOTSUPP

EPFNOSUPPORT

EAFNOSUPPORT

EADDRINUSE

EADDRNOTAVAIL

ENETDOWN

ENETUNREACH

ENETRESET

ECONNABORTED

ECONNRESET

ENOBUFS

EISCONN

ENOTCONN

ESHUTDOWN

ETIMEDOUT

ECONNREFUSED

ELOOP

ENAMETOOLONG

EHOSTDOWN

EHOSTUNREACH

ENOTEMPTY

EPROCLIM

EUSERS

EDQUOT

ESTALE

EBADRPC

ERPCMISMATCH

EPROGUNAVAIL

EPROGMISMATCH

EPROCUNAVAIL

ENOLCK

ENOSYS

EFTYPE

EAUTH

ENEEDAUTH

EIDRM

ENOMSG

EOVERFLOW

ECANCELED

EILSEQ

ENOATTR

EDOOFUS

EBADMSG

EMULTIHOP

ENOLINK

EPROTO

ENOTCAPABLE

ECAPMODE

ENOTRECOVERABLE

EOWNERDEAD

Methods

impl BsdError

fn from_errno() -> Option<BsdError>

Create BsdError from errno in current thread. Returns None of error is not known or there is no error at all.